requirements.txt
kafka-python
cassandra-driver
flask
requests
sseclient